+++ libs/exceptions/src/lib/exceptions/conflict-entity-creation.exception.ts
@@ -8,6 +8,7 @@ export class ConflictEntityCreationException extends AbstractHttpException<Confl
     public readonly entityName: string,
     public readonly fieldName: string,
     public readonly fieldValue: unknown,
+    errorCode?: string,
     rootCause?: unknown,
   ) {
     super(
@@ -15,6 +16,7 @@ export class ConflictEntityCreationException extends AbstractHttpException<Confl
       i18nString('exception.CONFLICT.CAN_NOT_CREATE_ENTITY'),
       HttpStatus.CONFLICT,
       { entityName, fieldName, fieldValue },
+      errorCode,
       rootCause,
     );
   }