+++ libs/exceptions/src/lib/swagger/properties/error-code-swagger.property.ts
@@ -1,0 +1,10 @@
+import { SchemaObject } from '@nestjs/swagger/dist/interfaces/open-api-spec.interface';
+
+export const errorCodeSwaggerProperty = (...errorCodes: string[]) => {
+  return {
+    errorCode: {
+      type: 'enum',
+      enum: errorCodes,
+    } as SchemaObject,
+  };
+};